switch to interface injection to make better unit tests
