+++ backend/src/main/java/com/bookbae/server/service/DatabasePoolServiceImpl.java
@@ -1,13 +1,18 @@
+package com.bookbae.server.service;
+
 import com.bookbae.server.DatabasePoolService;
 import org.apache.commons.dbcp2.BasicDataSource;
 import java.sql.Connection;
 import java.sql.SQLException;
+import jakarta.enterprise.context.ApplicationScoped;
 
-public class MockDatabaseService implements DatabasePoolService {
+@ApplicationScoped
+public class DatabasePoolServiceImpl implements DatabasePoolService {
     private BasicDataSource dataSource;
-    public MockDatabaseService() {
+
+    public DatabasePoolServiceImpl() {
         this.dataSource = new BasicDataSource();
-        this.dataSource.setUrl("jdbc:h2:mem:tests");
+        this.dataSource.setUrl("jdbc:h2:mem:db1");
         this.dataSource.setDriverClassName("org.h2.Driver");
         this.dataSource.setPoolPreparedStatements(true);
     }
