+++ backend/src/test/java/SQLFailService.java
@@ -1,10 +1,8 @@
-import com.bookbae.server.RestApplication;
+import com.bookbae.server.DatabasePoolService;
 import java.sql.Connection;
 import java.sql.SQLException;
 
-public class SQLFailService extends DatabasePoolService {
-    public SQLFailService() {}
-
+public class SQLFailService implements DatabasePoolService {
     @Override
     public Connection getConnection() throws SQLException {
         throw new SQLException();
